Move oidc to a new settings sub-command
